Update README.md
Working with Logic Errors

Logic Errors—errors in the logic of the algorithm. These errors emphasize the need for a correct algorithm. If the statements are out of order, if there are errors in a formula, or if there are missing steps, the program can still run and give you output, but it may be the wrong output. Since there is no list of errors for logic errors, you may not realize you have errors unless you check your output.

It is very important to know what output you expect. You should test your programs with different inputs, and know what output to expect in each case. Calculate each case by hand before running your program so that you know what to expect. You may get a correct answer for one case, but not for another case. This will help you figure out where your logic errors are.

Compile and run this program. You should get no syntax errors. Is the output correct? If not, why?

This program has no syntax or runtime errors, but it certainly has a logic error. Logic errors may create a challenge for the programmer. Find the problem.

Fix the bug. Ask for help if needed.